crud detailsPlan

+++ app/Http/Controllers/Admin/PlanController.php
@@ -52,13 +52,23 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                    ->with('details')
+                    ->where('url', $url)
+                    ->first();
 
         if (!$plan)
             return redirect()->back();
 
+        if($plan->details->count() > 0){
+            return redirect()->back()
+                            ->with('text-danger', 'Não foi possivel deletar, Plano com detalhes associados!!');
+
+        }
+
         $plan->delete();
-        return redirect()->route('plans.index');
+        return redirect()->route('plans.index')
+                        ->with('text-success', "Plano {$plan->name} foi deletado!!");
     }
 
     public function search(Request $request)
